+++ workflows_definition/RELEASES.md
@@ -75,3 +75,11 @@ Based on:
 - Speakeasy CLI 1.376.0 (2.402.5) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.4.0] workflows_definition
+
+## 2024-08-30 00:26:50
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/workflows-definition.yaml
+- Speakeasy CLI 1.383.2 (2.407.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.4.1] workflows_definition